backend/requirements.txt
google-generativeai
python-dotenv
PyPDF2
openai-agents
fastapi
uvicorn
python-multipart
faiss-cpu
langchain
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers
tiktoken
sqlalchemy
bcrypt
python-jose[cryptography]
pydantic[email]